chore: rename folder GeoMap back to Geo
