+++ giraffe/src/components/Geo/SVGIcon.ts
