+++ giraffe/src/components/Geo/bing-maps/Bing.ts
