+++ giraffe/src/components/Geo/bing-maps/BingLayerObject.ts
