+++ giraffe/src/components/Geo/bing-maps/BingMap.tsx
