+++ giraffe/src/components/Geo/dimensionCalculations.ts
