+++ giraffe/src/components/Geo/processing/tableProcessing.ts
